🐛 Fix:#34 기존 유저 정의 타입을 Prisma client 타입으로 수정
