+++ src/containers/work.container.ts
@@ -4,7 +4,7 @@ import { WorkRepository } from '#repositories/work.repository.js';
 import { WorkService } from '#services/work.service.js';
 
 // 여기서 Repository, Service, Controller를 연결합니다. 그리고 컨트롤러를 export 해줍니다.
-const workRepository = new WorkRepository(prismaClient);
+const workRepository = new WorkRepository(prismaClient.challengeWork);
 const workService = new WorkService(workRepository);
 const workController = new WorkController(workService);
