+++ src/repositories/feedback.repository.ts
@@ -1,15 +1,11 @@
+import type { Feedback, PrismaClient } from '@prisma/client';
 import type { IFeedbackRepository } from '#interfaces/repositories/feedback.repository.interface.js';
-import type { CreateFeedbackDTO, Feedback, UpdateFeedbackDTO } from '#types/work.types.js';
-import type { PrismaClient } from '@prisma/client';
+import type { CreateFeedbackDTO, UpdateFeedbackDTO } from '#types/feedback.types.js';
 
 export class FeedbackRepository implements IFeedbackRepository {
-  private feedback: PrismaClient['feedback'];
-  
-  constructor(client: PrismaClient) {
-    this.feedback = client.feedback; // 이 부분에 각 모델(스키마)를 연결합니다.
-  }
+  constructor(private feedback: PrismaClient['feedback']) {}
 
-    // 이 아래로 직접 DB와 통신하는 코드를 작성합니다.
+  // 이 아래로 직접 DB와 통신하는 코드를 작성합니다.
   // 여기서 DB와 통신해 받아온 데이터를 위로(service로) 올려줍니다.
   findMany = async (options: { orderBy: string; page: number; pageSize: number }): Promise<Feedback[] | null> => {
     const feedbacks = await this.feedback.findMany();
