+++ src/types/feedback.types.ts
@@ -1,0 +1,11 @@
+export interface CreateFeedbackDTO {
+  content: string;
+  image: string;
+  ownerId: string;
+  workId: string;
+}
+
+export interface UpdateFeedbackDTO {
+  content?: string;
+  image?: string;
+}
